ADM-443 change deploymentFrequent value type from double to float on backend

+++ backend/src/main/java/heartbeat/controller/report/dto/response/AvgDeploymentFrequency.java
@@ -17,7 +17,7 @@ public class AvgDeploymentFrequency {
 
 	private String deploymentFrequency;
 
-	public AvgDeploymentFrequency(Double deploymentFrequency) {
+	public AvgDeploymentFrequency(float deploymentFrequency) {
 		DecimalFormat df = new DecimalFormat(Constants.FORMAT_DOUBLE_2_DECIMALS);
 		this.deploymentFrequency = df.format(deploymentFrequency);
 	}
